simplified getTokens function and deleted generic files
