+++ apps/swirl-docs/src/components/Tokens/BorderTokens.tsx
@@ -1,5 +1,9 @@
-import { BorderTokenCategory, Token } from "@swirl/lib/tokens";
-import { getBorderTokens } from "@swirl/lib/tokens/src/borderTokens";
+import {
+  BorderTokenCategories,
+  BorderTokenCategory,
+  Token,
+} from "@swirl/lib/tokens";
+import { getTokens } from "@swirl/lib/tokens/src/utils";
 import { FunctionComponent } from "react";
 import TokensList from "./TokensList";
 
@@ -10,7 +14,13 @@ interface BorderTokenProps {
 export const BorderTokens: FunctionComponent<BorderTokenProps> = ({
   category,
 }) => {
-  const categories = getBorderTokens();
+  const categories = getTokens(
+    {
+      borderRadius: [],
+      borderWidth: [],
+    },
+    (type: string) => BorderTokenCategories.includes(type)
+  );
 
   const tokens: Token[] = categories[category] ?? [];
