+++ apps/swirl-docs/src/components/Tokens/SpacingTokens.tsx
@@ -1,7 +1,15 @@
-import { getSpacingTokens } from "@swirl/lib/tokens/src/spacingTokens";
+import { SpacingTokenCategories } from "@swirl/lib/tokens";
+import { getTokens } from "@swirl/lib/tokens/src/utils";
 import { FunctionComponent } from "react";
 import TokensList from "./TokensList";
 
 export const SpacingTokens: FunctionComponent = () => {
-  return <TokensList tokens={getSpacingTokens().spacing} />;
+  const tokens = getTokens(
+    {
+      spacing: [],
+    },
+    (type: string) => SpacingTokenCategories.includes(type)
+  );
+
+  return <TokensList tokens={tokens.spacing} />;
 };
