+++ apps/swirl-docs/src/components/Tokens/TypographyTokens.tsx
@@ -1,5 +1,9 @@
-import { Token, TypographyTokenCategory } from "@swirl/lib/tokens";
-import { getTypographyTokens } from "@swirl/lib/tokens/src/typographyTokens";
+import {
+  Token,
+  TypographyTokenCategories,
+  TypographyTokenCategory,
+} from "@swirl/lib/tokens";
+import { getTokens } from "@swirl/lib/tokens/src/utils";
 import { FunctionComponent } from "react";
 import TokensList from "./TokensList";
 
@@ -10,7 +14,16 @@ interface TypographyTokensProps {
 export const TypographyTokens: FunctionComponent<TypographyTokensProps> = ({
   typographyCategory,
 }) => {
-  const category = getTypographyTokens();
+  const category = getTokens(
+    {
+      fontWeights: [],
+      lineHeights: [],
+      letterSpacing: [],
+      fontFamily: [],
+      fontSizes: [],
+    },
+    (type: string) => TypographyTokenCategories.includes(type)
+  );
 
   const tokens: Token[] = category[typographyCategory] ?? [];
